refactor: Change data type

+++ rules.c
@@ -12,9 +12,9 @@
 
 #include "push_swap.h"
 
-void	swap(int *vector, char stack)
+void	swap(long long *vector, char stack)
 {
-	int	tmp;
+	long long	tmp;
 
 	tmp = vector[0];
 	vector[0] = vector[1];
@@ -26,13 +26,13 @@ void	swap(int *vector, char stack)
 }
 
 // 212 32 5 42 9
-void	rotate(int *vector, int size, char stack)
+void	rotate(long long *vector, int size, char stack)
 {
-	int	tmp;
+	long long	tmp;
 
 	size -= 1;
 	tmp = vector[0];
-	ft_memmove(&vector[0], &vector[1], sizeof(int) * (size));
+	ft_memmove(&vector[0], &vector[1], sizeof(long long) * (size));
 	vector[size] = tmp;
 	if (stack == 'a')
 		write(1, "ra\n", 3);
@@ -40,12 +40,12 @@ void	rotate(int *vector, int size, char stack)
 		write(1, "rb\n", 3);
 }
 
-void	reverse_rotate(int *vector, int size, char stack)
+void	reverse_rotate(long long *vector, int size, char stack)
 {
-	int	tmp;
+	long long	tmp;
 
 	tmp = vector[size - 1];
-	ft_memmove(&vector[1], &vector[0], sizeof(int) * (size - 1));
+	ft_memmove(&vector[1], &vector[0], sizeof(long long) * (size - 1));
 	vector[0] = tmp;
 	if (stack == 'a')
 		write(1, "rra\n", 4);
@@ -53,15 +53,15 @@ void	reverse_rotate(int *vector, int size, char stack)
 		write(1, "rrb\n", 4);
 }
 
-void	push(int *dst, int *src, t_stack *stack, char move)
+void	push(long long *dst, long long *src, t_stack *stack, char move)
 {
-	int	tmp;
+	long long	tmp;
 
 	tmp = src[0];
 	if (stack->size_a > 0)
 	{
-		ft_memmove(src, &src[1], (stack->src_len - 1) * sizeof(int));
-		ft_memmove(&dst[1], dst, (stack->dst_len) * sizeof(int));
+		ft_memmove(src, &src[1], (stack->src_len - 1) * sizeof(long long));
+		ft_memmove(&dst[1], dst, (stack->dst_len) * sizeof(long long));
 		dst[0] = tmp;
 	}
 	if (move == 'a')
